Refactored staticmethods from measurement inputs.
