+++ src/cascade_at/settings/settings.py
@@ -1,6 +1,6 @@
 import json
 
-from cascade_at.settings.settings_configuration import SettingsConfiguration
+from cascade_at.settings.settings_config import SettingsConfig
 from cascade_at.core.db import db_tools
 from cascade_at.core.errors import SettingsError
 from cascade_at.core.log import get_loggers
@@ -22,7 +22,7 @@ def load_settings(settings_json):
     Returns:
         cascade_at.settings.configuration.Configuration
     """
-    settings = SettingsConfiguration(settings_json)
+    settings = SettingsConfig(settings_json)
     errors = settings.validate_and_normalize()
     if errors:
         print(f"Configuration does not validate {errors}")
